+++ problems/47.permutations-ii.cpp
@@ -1,0 +1,12 @@
+// @leetcode id=47 questionId=47 slug=permutations-ii lang=cpp site=leetcode.com title="Permutations II"
+class Solution {
+public:
+    vector<vector<int>> permuteUnique(vector<int>& nums) {
+        vector<vector<int>> result;
+        sort(nums.begin(), nums.end());
+        do {
+            result.push_back(nums);
+        } while (next_permutation(nums.begin(), nums.end()));
+        return result;
+    }
+};